refactor: rename module, update and split subclasses
