--- wrappers/__init__.py
@@ -1,1 +1,0 @@
-# Empty __init__.py file
